requirements.txt
asgiref==3.8.1
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-axes==8.0.0
django-debug-toolbar==5.2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-js-asset==3.1.2
django-recaptcha==4.1.0
django-tinymce==4.1.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
idna==3.10
loguru==0.7.3
mysql-connector-python==9.3.0
openpyxl==3.1.5
pillow==11.2.1
pymemcache==4.0.0
python-dotenv==1.1.0
python-memcached==1.62
pytz==2025.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
win32_setctime==1.2.0
